Removed all the warnings and deprecated
