+++ NtgeCore-Android/.gitignore
