+++ NtgeCore-Android/build.gradle.kts
