+++ NtgeCore-Android/gradle.properties
