+++ NtgeCore-Android/settings.gradle.kts
