+++ NtgeCore-Android/src/main/AndroidManifest.xml
