+++ NtgeCore-Android/src/main/java/com/dimension/ntge/Ntge.kt
