+++ NtgeCore-Android/src/main/res/values/strings.xml
