+++ ntge-core/src/ed25519/keypair.rs
@@ -1,4 +1,4 @@
-pub use ed25519_dalek::{self, PublicKey, SecretKey};
+pub use ed25519_dalek::{self, Keypair, PublicKey, SecretKey};
 use rand::rngs::OsRng;
 
 use crate::{ed25519::private::Ed25519PrivateKey, ed25519::public::Ed25519PublicKey};
@@ -32,13 +32,8 @@ impl Ed25519Keypair {
     }
 
     pub fn construct_from_private_key(private_key: &Ed25519PrivateKey) -> Self {
-        let sk: SecretKey = SecretKey::from_bytes(&(private_key.raw.to_bytes())).unwrap();
-        let pk: PublicKey = (&sk).into();
-
-        let keypair = ed25519_dalek::Keypair {
-            secret: sk,
-            public: pk,
-        };
+        let secret_key: &SecretKey = &private_key.raw;
+        let keypair: Keypair = construct_from_secret_key(&secret_key);
 
         Ed25519Keypair { raw: keypair }
     }
@@ -49,6 +44,22 @@ impl Drop for Ed25519Keypair {
         if cfg!(feature = "drop-log-enable") {
             println!("{:?} is being deallocated", self);
         }
+    }
+}
+
+pub fn create_keypair() -> Keypair {
+    // a.k.a Cryptographically secure pseudo-random number generator.
+    let mut csprng: OsRng = OsRng {};
+    ed25519_dalek::Keypair::generate(&mut csprng)
+}
+
+pub fn construct_from_secret_key(private_key: &SecretKey) -> Keypair {
+    let sk: SecretKey = SecretKey::from_bytes(&(private_key.to_bytes())).unwrap();
+    let pk: PublicKey = (&sk).into();
+
+    Keypair {
+        public: pk,
+        secret: sk,
     }
 }
 
@@ -87,4 +98,36 @@ pub extern "C" fn c_ed25519_keypair_construct_from_private_key(
 ) -> *mut Ed25519Keypair {
     let keypair = Ed25519Keypair::construct_from_private_key(&private_key);
     Box::into_raw(Box::new(keypair))
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::ed25519;
+
+    #[test]
+    fn it_creates_a_keypair() {
+        let keypair = ed25519::keypair::create_keypair();
+
+        let serialized_private_key = ed25519::private::serialize_private_key(&(keypair.secret));
+        println!("{}", serialized_private_key);
+
+        let serialized_public_key = ed25519::public::serialize_public_key(&(keypair.public));
+        println!("{}", serialized_public_key);
+
+        let deserialized_private_key =
+            ed25519::private::deserialize_private_key(&serialized_private_key);
+        assert_eq!(deserialized_private_key.is_ok(), true);
+        assert_eq!(
+            keypair.secret.to_bytes(),
+            deserialized_private_key.unwrap().to_bytes()
+        );
+
+        let deserialized_public_key =
+            ed25519::public::deserialize_public_key(&serialized_public_key);
+        assert_eq!(deserialized_public_key.is_ok(), true);
+        assert_eq!(
+            keypair.public.to_bytes(),
+            deserialized_public_key.unwrap().to_bytes()
+        );
+    }
 }
