+++ ntge-core/src/key_utils.rs
@@ -18,7 +18,7 @@ pub fn keypair_validation(
     private_key: &SecretKey,
     public_key: &PublicKey,
 ) -> Result<bool, error::CoreError> {
-    let keypair: Keypair = construct_from_private_key(private_key);
+    let keypair: Keypair = keypair::construct_from_secret_key(private_key);
     if keypair.public.to_bytes() != (*public_key).to_bytes() {
         let core_error = error::CoreError::KeyInvalidError {
             name: CURVE_NAME_ED25519,
@@ -87,4 +87,80 @@ pub unsafe extern "C" fn c_key_utils_ed25519_private_key_to_x25519(
         raw: ed25519_private_key_to_x25519(&ed25519_private_key.raw),
     };
     Box::into_raw(Box::new(x25519_private_key))
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{ed25519, key_utils};
+    use rand::rngs::OsRng;
+    use x25519_dalek::{EphemeralSecret, PublicKey, StaticSecret};
+
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
+    }
+
+    #[test]
+    fn it_validates_a_valid_keypair() {
+        let keypair = ed25519::keypair::create_keypair();
+        assert!(key_utils::keypair_validation(&keypair.secret, &keypair.public).is_ok());
+    }
+
+    #[test]
+    fn it_validates_an_invalid_keypair() {
+        let keypair1 = ed25519::keypair::create_keypair();
+        let keypair2 = ed25519::keypair::create_keypair();
+
+        assert!(!(key_utils::keypair_validation(&keypair1.secret, &keypair2.public).is_ok()));
+    }
+
+    #[test]
+    fn it_converts_an_ed25519_public_key_to_x25519() {
+        let keypair = ed25519::keypair::create_keypair();
+        let pubkey = key_utils::ed25519_public_key_to_x25519(&keypair.public);
+
+        // no need to verify an X25519 key
+        assert_ne!(pubkey.as_bytes(), &[0; 32]);
+    }
+
+    #[test]
+    fn it_converts_an_ed25519_private_key_to_x25519() {
+        let keypair = ed25519::keypair::create_keypair();
+        let private_key = key_utils::ed25519_private_key_to_x25519(&keypair.secret);
+
+        assert_ne!(private_key.to_bytes(), [0; 32]);
+    }
+
+    #[test]
+    fn it_conducts_diffie_hellman_on_two_ed25519_keypairs() {
+        let alice = ed25519::keypair::create_keypair();
+        let alice_private = key_utils::ed25519_private_key_to_x25519(&alice.secret);
+        let alice_public = key_utils::ed25519_public_key_to_x25519(&alice.public);
+
+        let bob = ed25519::keypair::create_keypair();
+        let bob_private = key_utils::ed25519_private_key_to_x25519(&bob.secret);
+        let bob_public = key_utils::ed25519_public_key_to_x25519(&bob.public);
+
+        //Alice -> Bob
+        let alice_private_ephemeral = EphemeralSecret::new(&mut OsRng);
+        let alice_public_ephemeral = PublicKey::from(&alice_private_ephemeral);
+        let alice_shared_secret1 = alice_private_ephemeral.diffie_hellman(&bob_public);
+        let bob_private_static: StaticSecret = StaticSecret::from(bob_private.to_bytes());
+        let bob_shared_secret1 = bob_private_static.diffie_hellman(&alice_public_ephemeral);
+        assert_eq!(
+            alice_shared_secret1.as_bytes(),
+            bob_shared_secret1.as_bytes()
+        );
+
+        //Bob -> Alice
+        let bob_private_ephemeral = EphemeralSecret::new(&mut OsRng);
+        let bob_public_ephemeral = PublicKey::from(&bob_private_ephemeral);
+        let bob_shared_secret2 = bob_private_ephemeral.diffie_hellman(&alice_public);
+        let alice_private_static: StaticSecret = StaticSecret::from(alice_private.to_bytes());
+        let alice_shared_secret2 = alice_private_static.diffie_hellman(&bob_public_ephemeral);
+        assert_eq!(
+            alice_shared_secret2.as_bytes(),
+            bob_shared_secret2.as_bytes()
+        );
+    }
 }
