+++ ntge-core/src/x25519/mod.rs
@@ -4,12 +4,3 @@
 pub mod filekey;
 pub mod private;
 pub mod public;
-
-use hkdf::Hkdf;
-use rand::rngs::OsRng;
-use rand::RngCore;
-use secrecy::{ExposeSecret, Secret};
-use sha2::Sha256;
-use x25519_dalek::{EphemeralSecret, PublicKey, StaticSecret};
-
-use crate::{aead, error::CoreError, message::recipient::MessageRecipientHeader};
