+++ ntge/src/bin/ntge/encrypt.rs
@@ -1,5 +1,5 @@
-use ntge_core::ed25519::public::Ed25519PublicKey;
-use ntge_core::key_utils;
+
+
 use ntge_core::message::{encryptor::Encryptor, Message};
 use ntge_core::x25519::public::X25519PublicKey;
